+++ src/colorpage.cpp
@@ -40,7 +40,7 @@ void ColorPage::ambientUpdateReceived(std::uint32_t newAmbientValue, std::uint32
     emit ambientUpdate(newAmbientValue, newBrightness);
 }
 
-void ColorPage::show(const QColor& color, std::uint32_t lightCount, EColorPickerType bestType) {
+void ColorPage::update(const QColor& color, std::size_t lightCount, EColorPickerType bestType) {
     mColor = color;
     mBestType = bestType;
     if (lightCount == 0) {