+++ src/colorpage.h
@@ -46,8 +46,8 @@ public:
     /// programmatically updates brightness
     void updateBrightness(std::uint32_t brightness);
 
-    /// called when the widget is shown
-    void show(const QColor& color, std::uint32_t lightCount, EColorPickerType bestType);
+    /// called when the app state is updated (IE when the selected lights changed)
+    void update(const QColor& color, std::size_t lightCount, EColorPickerType bestType);
 
     /// getter for currently selected color
     const QColor color() const noexcept { return mColor; }