+++ src/editablefieldwidget.cpp
@@ -9,6 +9,7 @@
 #include <QMessageBox>
 
 #include "utils/reachability.h"
+#include "utils/qt.h"
 
 EditableFieldWidget::EditableFieldWidget(const QString& text,
                                          QWidget* parent,
@@ -24,20 +25,20 @@ EditableFieldWidget::EditableFieldWidget(const QString& text,
     mEditableField = new QLineEdit(text, this);
     mEditableField->setFont(mText->font());
     mEditableField->setVisible(false);
-    mEditableField->setSizePolicy(QSizePolicy::Expanding, QSizePolicy::Expanding);
+    mEditableField->setSizePolicy(QSizePolicy::Fixed, QSizePolicy::Fixed);
     mEditableField->setStyleSheet("border-style: outset; border-width: 1px; border-color: white; "
                                   "background-color:rgb(30,30,30);");
     mEditableField->setAttribute(Qt::WA_TransparentForMouseEvents, false);
     connect(mEditableField, SIGNAL(textEdited(QString)), this, SLOT(lineEditChanged(QString)));
 
     mRightButton = new QPushButton(this);
-    mRightButton->setSizePolicy(QSizePolicy::Expanding, QSizePolicy::Expanding);
+    mRightButton->setSizePolicy(QSizePolicy::Fixed, QSizePolicy::Fixed);
     mRightButton->setAttribute(Qt::WA_TransparentForMouseEvents, false);
     mRightButton->setStyleSheet("border: none;");
     connect(mRightButton, SIGNAL(clicked(bool)), this, SLOT(rightButtonClicked(bool)));
 
     mLeftButton = new QPushButton(this);
-    mLeftButton->setSizePolicy(QSizePolicy::Expanding, QSizePolicy::Expanding);
+    mLeftButton->setSizePolicy(QSizePolicy::Fixed, QSizePolicy::Fixed);
     mLeftButton->setAttribute(Qt::WA_TransparentForMouseEvents, false);
     mLeftButton->setStyleSheet("border: none;");
     mLeftButton->setVisible(false);
@@ -48,13 +49,6 @@ EditableFieldWidget::EditableFieldWidget(const QString& text,
 
     mRightButton->setIcon(QIcon(mEditIcon));
     mLeftButton->setIcon(QIcon(QPixmap(":/images/closeX.png")));
-
-    mLayout = new QHBoxLayout(this);
-    mLayout->addWidget(mText, 16);
-    mLayout->addWidget(mLeftButton, 2);
-    mLayout->addWidget(mRightButton, 2);
-    mLayout->setContentsMargins(0, 0, 0, 0);
-    mLayout->setSpacing(0);
 
     mMaxFieldSize = maxFieldSize;
     mMaxFieldError = maxFieldError;
@@ -126,14 +120,12 @@ void EditableFieldWidget::setInEditMode(bool inEditMode) {
     mIsEditing = inEditMode;
     if (inEditMode) {
         // turn on editing
-        mLayout->replaceWidget(mText, mEditableField);
         mEditableField->setVisible(true);
         mText->setVisible(false);
         mLeftButton->setVisible(true);
         mRightButton->setIcon(mCheckIcon);
     } else {
         // turn off editing and signal it
-        mLayout->replaceWidget(mEditableField, mText);
         mEditableField->setVisible(false);
         mText->setVisible(true);
         mLeftButton->setVisible(false);
@@ -156,4 +148,25 @@ void EditableFieldWidget::enableEditing(bool enableEditing) {
 
 void EditableFieldWidget::requireIPAddress(bool requireIP) {
     mRequireIP = requireIP;
+}
+
+void EditableFieldWidget::resizeEvent(QResizeEvent*) {
+    resize();
+}
+
+void EditableFieldWidget::resize() {
+    auto h = height();
+    auto buttonSize = h * 0.8;
+    auto w = width();
+    mRightButton->setGeometry(w - buttonSize, buttonSize * 0.1, buttonSize, buttonSize);
+    mLeftButton->setGeometry(w - 2 * buttonSize, buttonSize * 0.1, buttonSize, buttonSize);
+    if (mIsEditing) {
+        cor::resizeIcon(mRightButton, ":/images/checkmark.png");
+    } else {
+        cor::resizeIcon(mRightButton, ":/images/editIcon.png");
+    }
+    cor::resizeIcon(mLeftButton, ":/images/closeX.png");
+
+    mText->setGeometry(0, 0, w - 2 * buttonSize, h);
+    mEditableField->setGeometry(mText->geometry());
 }